// lib/tax-forms/frontend-types.ts
export enum TaxFormType {
  W9 = 'W9',
  W8_BEN = 'W8_BEN',
  W8_BEN_E = 'W8_BEN_E',
}

export enum SubmitterType {
  Individual = 'Individual',
  Business = 'Business',
}

export enum FederalTaxClassification {
  Individual = 'Individual',
  C_Corporation = 'C_Corporation',
  S_Corporation = 'S_Corporation',
  Partnership = 'Partnership',
  TrustEstate = 'TrustEstate',
  LimitedLiabilityCompany = 'LimitedLiabilityCompany',
  Other = 'Other',
}
// ...
export type W9TaxFormValues = {
  isUSPersonOrEntity?: boolean;
  submitterType?: SubmitterType;
  formType?: TaxFormType;
  email?: string;
  signer?: { firstName?: string; middleName?: string; lastName?: string };
  isSigned?: boolean;
  businessName?: string;
  federalTaxClassificationDetails?: string;
  exemptPayeeCode?: string;
  fatcaExemptionCode?: string;
  taxIdNumberType?: 'SSN' | 'EIN';
  taxIdNumber?: string;
  accountNumbers?: string;
  federalTaxClassification?: FederalTaxClassification;
  hasConfirmedTOS?: boolean;
  location?: {
    country?: string;
    structured?: { address1?: string; address2?: string; city?: string; zone?: string; postalCode?: string };
  };
};